Move shared map flow geometry into lib for analytics imports.
Stop homeworld locator analytics from depending on map-graph components
for pure coordinate helpers used by attention and context menus.
