+++ packages/frontend/src/lib/mapFlowGeometry.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { flowPointNeedsPan } from './geometry'
+import { flowPointNeedsPan } from './mapFlowGeometry'
 
 describe('flowPointNeedsPan', () => {
   const centered = {
